40. 멀티쓰레드 로딩 추가

+++ Engine/Include/Component/RadioButton_Com.cpp
@@ -77,7 +77,7 @@ int RadioButton_Com::Update(float DeltaTime)
 	{
 		if (m_vecRadio[i]->CheckCom->GetCheckState() == CBS_TRUE)
 		{
-			SelectIndex = i;
+			SelectIndex = (int)i;
 
 			if (SelectIndex < PrevSelect)
 				break;
@@ -95,6 +95,14 @@ int RadioButton_Com::Update(float DeltaTime)
 		PrevSelect = SelectIndex;
 		SelectIndex = INT_MAX;
 	}
+
+	//설명
+	//true가 발생하면 일단 인덱스를 저장한다.
+	//저장 후 밑에서 선택된 인덱스를 TRUE로 만든다.
+	//거기서 PrevIndex를 저장한다.
+	//다시 위로가서 PrevIndex와 현재인덱스를 비교하여 continue인지 break인지 판단
+	//PrevIndex번째를 FALSE로 만든다.
+	//continue시 ->방향체크 break시 <-방향체크.
 
 	return 0;
 }
